# public/docs/openApi2.json
{
  "swagger": "2.0",
  "info": {
    "contact": {
      "email": "rwalle61@yahoo.com"
    },
    "description": "API to track league results",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "title": "League Tracking API",
    "version": "1.0.0"
  },
  "host": "leaguetracker.appspot.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/seasons": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "season creation options",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "playersOptions": {
                  "items": {
                    "$ref": "#/definitions/PlayerOptions"
                  },
                  "type": "array"
                },
                "seasonName": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Season created",
            "schema": {
              "properties": {
                "players": {
                  "items": {
                    "$ref": "#/definitions/Player"
                  },
                  "type": "array"
                },
                "seasonName": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "400": {
            "description": "Bad input parameter"
          }
        },
        "tags": [
          "users"
        ],
        "summary": "Create a new season"
      },
      "put": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "season update options",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "games": {
                  "items": {
                    "properties": {
                      "namesOfLosers": {
                        "items": {
                          "type": "string"
                        },
                        "type": "array"
                      },
                      "namesOfWinners": {
                        "items": {
                          "type": "string"
                        },
                        "type": "array"
                      }
                    },
                    "type": "object"
                  },
                  "type": "array"
                },
                "season": {
                  "properties": {
                    "players": {
                      "items": {
                        "$ref": "#/definitions/Player"
                      },
                      "type": "array"
                    },
                    "seasonName": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              },
              "required": [
                "season",
                "games"
              ],
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Season updated",
            "schema": {
              "properties": {
                "deltas": {
                  "items": {
                    "type": "integer"
                  },
                  "type": "array"
                },
                "season": {
                  "properties": {
                    "players": {
                      "items": {
                        "$ref": "#/definitions/Player"
                      },
                      "type": "array"
                    },
                    "seasonName": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              },
              "type": "object"
            }
          },
          "400": {
            "description": "Bad input parameter"
          }
        },
        "tags": [
          "users"
        ],
        "summary": "Update a season"
      }
    }
  },
  "definitions": {
    "ID": {
      "type": "string"
    },
    "Player": {
      "properties": {
        "id": {
          "$ref": "#/definitions/ID"
        },
        "losses": {
          "example": 1,
          "type": "integer"
        },
        "name": {
          "example": "Craig Forrest",
          "type": "string"
        },
        "score": {
          "example": 990,
          "type": "integer"
        },
        "wins": {
          "example": 0,
          "type": "integer"
        }
      },
      "required": [
        "name",
        "score",
        "wins",
        "losses"
      ],
      "type": "object"
    },
    "PlayerOptions": {
      "properties": {
        "id": {
          "$ref": "#/definitions/ID"
        },
        "losses": {
          "example": 1,
          "type": "integer"
        },
        "name": {
          "example": "Craig Forrest",
          "type": "string"
        },
        "score": {
          "example": 990,
          "type": "integer"
        },
        "wins": {
          "example": 0,
          "type": "integer"
        }
      },
      "required": [
        "name"
      ],
      "type": "object"
    }
  },
...
}